Consequencify sinewave lasers and Monsieur Brackeis' movement
Also create him a prefab
