+++ Assets/Prefabs/Enemies/JeanBrackeis.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1c22895df00a8bd40822f080834f087c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
